lab8/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.2.0
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
bcrypt==4.0.1
passlib==1.7.4
python-jose==3.3.0
email-validator==2.1.0
pymongo==4.6.1
dnspython==2.4.2
redis==5.0.1
pytest==7.4.2
pytest-asyncio==0.21.1
httpx==0.25.2